new_baseline.py cv:0.8090939027883142 public:0.8613

+++ new_baseline.py
@@ -19,9 +19,11 @@ from sklearn.decomposition import PCA
 from sklearn.preprocessing import LabelEncoder
 from sklearn.metrics import mean_squared_error
 from sklearn.metrics import mean_squared_log_error
+from sklearn.neighbors import KNeighborsRegressor
 from sklearn.model_selection import StratifiedKFold
 from sklearn.model_selection import train_test_split
 from sklearn.feature_extraction.text import CountVectorizer
+
 
 import warnings
 warnings.simplefilter('ignore')
@@ -32,7 +34,6 @@ RANDOM_SEED = 72
 train = pd.read_csv('./features/train.csv')
 test  = pd.read_csv('./features/test.csv')
 
-# TODO: ここ最後に回して、カラム削除もやってしまえばいいのでは
 def target_encoding(train, test, target_col, y_col, validation_col):
     # 学習データ全体でカテゴリにおけるyの平均を計算
     data_tmp    = pd.DataFrame({'target': train[target_col], 'y': train[y_col]})
@@ -55,37 +56,14 @@ def target_encoding(train, test, target_col, y_col, validation_col):
 for target in ['Platform', 'Genre', 'Rating']:
     target_encoding(train, test, target, 'Global_Sales', 'Publisher')
 
-'''頭がバグったのでちょっと中止　そもそもリークするのか？ いや、もちろん可能性はあるけど...
-# trainにしかない特徴量 ['EU_Sales', 'Global_Sales', 'JP_Sales', 'NA_Sales', 'Other_Sales']を使う
-# リークの可能性があるので、target encodingの要領でfoldを分けて集計する
-# def make_sales_portfolio(train, test, target_col, y_col, validation_col):
-    # 学習データ全体でジャンルごとに正規化した各Salesの情報を得る
-    # data_tmp = pd.DataFrame({'Genre': train[target_col], 'y': train[y_col]})
-    # _df = data_tmp.groupby('Genre')['y'].agg(['mean', 'max', 'sum'])
-    # _df = _df.add_prefix(f'{y_col}_').add_suffix(f'_std_of_{target_col}') # カラム名意味わからないけどしゃあない
-    # _df = ((_df - _df.min()) / (_df.max() - _df.min())).reset_index() # ジャンルごとに正規化
-    # test = pd.merge(test, _df, on='Genre', how='left')
-    # # 変換後の値を格納する配列を準備
-    # tmp = np.repeat(np.nan, train.shape[0])
-    # skf = StratifiedKFold(n_splits=N_SPLITS, random_state=RANDOM_SEED, shuffle=True)
-    # for train_idx, test_idx in skf.split(train, train[validation_col]):
-    #     # 学習データに対して、各カテゴリにおける目的変数の平均を計算
-    #     # _df = data_tmp.iloc[train_idx].groupby('Genre')['y'].agg(['mean', 'max', 'sum'])
-    #     _df = data_tmp.iloc[train_idx].groupby('Genre')['y'].sum()
-    #     # バリデーションデータについて、変換後の値を一時配列に格納
-    #     tmp[test_idx] = train[target_col].iloc[test_idx].map(_df)
-    #     print(_df)
-    #     print(tmp, sum(tmp))
-    #     exit()
-# for y_col in ['EU_Sales', 'Global_Sales', 'JP_Sales', 'NA_Sales', 'Other_Sales']:
-#     make_sales_portfolio(train, test, 'Genre', y_col, 'Publisher')
-#     exit()
-'''
-
 # 処理をまとめてやるためにtrainとtestを結合
 train_length = len(train) # あとで分離するように長さを保存
 train_test   = pd.concat([train, test], ignore_index=True) # indexを再定義
 # train_test   = train_test.fillna('none')
+# train_test["Publisher"] = train_test["Publisher"].replace("Unknown", '')
+train_test['Name']   = train_test['Name'].fillna('No_Title') # NameがNaNのものがあるので'No_Title'に変換
+train_test['Genre']  = train_test['Genre'].fillna('none') # floatとstrの比較になるので置換
+train_test['Rating'] = train_test['Rating'].fillna('E') # 全年齢で埋める
 
 # NameのEmbeddingsをt-sneかけたものを特徴量として加える
 embeddings_tsne = np.load('./features/sentence_embeddings_tsne.npy')
@@ -105,9 +83,6 @@ def add_tbd(df):
     return df
 
 train_test = add_tbd(train_test)
-
-# 同じNameのが出てる->プラットフォームで売り上げが分散する可能性？ 'Name'の出現回数を数えて特徴量にする、ついでに他の特徴量もいくつかcount encoding
-train_test['Name'] = train_test['Name'].fillna('No_Title') # NameがNaNのものがあるので'No_Title'に変換
 
 # PlatformとGenreを単純に文字列として結合してCountEncodingする
 train_test['Platform_and_Genre'] = train_test['Platform'] + '_' + train_test['Genre']
@@ -133,24 +108,21 @@ def count_encoding(df, target_col):
     _df = pd.DataFrame(train_test[target_col].value_counts().reset_index()).rename(columns={'index': target_col, target_col: f'CE_{target_col}'})
     return pd.merge(df, _df, on=target_col, how='left')
 
-for target_col in ['Name','Year_of_Release','Platform','Genre','Platform_and_Genre','Platform_and_Genre_and_Binning_Year']:
-    train_test = count_encoding(train_test, target_col)
-
 def label_encoding(df, target_col):
     le = preprocessing.LabelEncoder()
     df[f'LE_{target_col}'] = le.fit_transform(df[target_col])
     return df
 
-train_test['Genre']  = train_test['Genre'].fillna('none') # floatとstrの比較になるので置換
-for target_col in ['Year_of_Release','Platform','Genre']:
-    train_test = label_encoding(train_test, target_col)
-
 def onehot_encoding(df, target_col):
     _df = pd.get_dummies(df[target_col], dummy_na=False).add_prefix(f'OH_{target_col}=')
     return pd.concat([train_test, _df], axis=1)
 
-for target_col in ['Year_of_Release','Platform','Genre']:
-    train_test = onehot_encoding(train_test, target_col)
+# 同じNameのが出てる->プラットフォームで売り上げが分散する可能性？ 'Name'の出現回数を数えて特徴量にする、ついでに他の特徴量もいくつかcount encoding
+for target_col in ['Name','Year_of_Release','Platform','Genre','Rating','Platform_and_Genre','Platform_and_Genre_and_Binning_Year']:
+    if target_col!='Name':
+        train_test = onehot_encoding(train_test, target_col)
+    train_test = count_encoding(train_test, target_col)
+    train_test = label_encoding(train_test, target_col)
 
 # プラットフォームでのジャンルごとの売り上げの平均、最大、最小、合計を計算してプラットフォームでのジャンルの特徴を捉える NOTE: カウントとか効きそう？ 各国ごとに特徴量を作るのは効くのか？
 for sales in ['EU_Sales','Global_Sales','JP_Sales','NA_Sales','Other_Sales','Global_Sales']:
@@ -163,6 +135,16 @@ for sales in ['EU_Sales','Global_Sales','JP_Sales','NA_Sales','Other_Sales','Glo
     _df = pd.DataFrame(train_test.groupby(['Genre'])[sales].agg(['mean', 'max', 'min', 'sum']).reset_index())
     _df = _df.rename(columns={'mean': f'Genre_{sales}_mean', 'max': f'Genre_{sales}_max', 'min': f'Genre_{sales}_min', 'sum': f'Genre_{sales}_sum'})
     train_test = pd.merge(train_test, _df, on='Genre', how='left')
+
+    # Rating
+    _df = pd.DataFrame(train_test.groupby(['Rating'])[sales].agg(['mean', 'max', 'min', 'sum']).reset_index())
+    _df = _df.rename(columns={'mean': f'Rating_{sales}_mean', 'max': f'Rating_{sales}_max', 'min': f'Rating_{sales}_min', 'sum': f'Rating_{sales}_sum'})
+    train_test = pd.merge(train_test, _df, on='Rating', how='left')
+
+    # Year_of_Release_fillna
+    _df = pd.DataFrame(train_test.groupby(['Year_of_Release_fillna'])[sales].agg(['mean', 'max', 'min', 'sum']).reset_index())
+    _df = _df.rename(columns={'mean': f'Year_of_Release_fillna_{sales}_mean', 'max': f'Year_of_Release_fillna_{sales}_max', 'min': f'Year_of_Release_fillna_{sales}_min', 'sum': f'Year_of_Release_fillna_{sales}_sum'})
+    train_test = pd.merge(train_test, _df, on='Year_of_Release_fillna', how='left')
 
 # 'Rating'の出現回数を数えて、よく売れそうな対象年齢について考える
 _df = pd.DataFrame(train_test.groupby(['Rating'])['Global_Sales'].agg(['mean', 'max', 'min']).reset_index())
@@ -220,6 +202,11 @@ train_test = pd.merge(train_test, _df, on='Name', how='left')
 
 train_test['total_sale_count'] = train_test['sale_count'] * train_test['resale_count']
 
+# ソフト数の変遷(同年に複数タイトル売られたら相対的に売り上げが減るかも)
+_df = pd.DataFrame(train_test.groupby(['Year_of_Release_fillna'])['Year_of_Release_fillna'].agg(['count']).reset_index())
+_df = _df.rename(columns={'count': 'published_game_count_in_year'})
+train_test = pd.merge(train_test, _df, on='Year_of_Release_fillna', how='left')
+
 # 年ごとのクチコミの多さ
 _df = pd.DataFrame(train_test.groupby(['Year_of_Release'])['User_Count'].agg(['count']).reset_index())
 _df = _df.rename(columns={'count': 'Year_of_Release_User_Count_count'})
@@ -240,7 +227,7 @@ train_test = pd.concat([train_test, _df], axis=1)
 # ここのPublisherの扱い方参考 https://www.guruguru.science/competitions/13/discussions/386fb2ed-f0a6-4706-85ba-7a03fedea375/
 for target_col in ['Platform','Genre','Year_of_Release']:
     # 各PublisherのPlatform毎のデータ件数は以下のようにして集計しています。
-    plat_pivot = train_test.pivot_table(index='Publisher', columns=target_col,values='Name', aggfunc='count').reset_index()
+    plat_pivot = train_test.pivot_table(index='Publisher', columns=target_col, values='Name', aggfunc='count').reset_index()
     plat_pivot = plat_pivot.fillna(0) # カウントだから0がいいはず
 
     # 行列の標準化
@@ -285,63 +272,6 @@ _df = pd.DataFrame(train_test.groupby(['Platform'])['User_Count'].agg(['mean', '
 _df = _df.add_prefix('Platform_User_Count_').rename(columns={'Platform_User_Count_Platform': 'Platform'})
 train_test = pd.merge(train_test, _df, on='Platform', how='left')
 
-
-'''近傍を使う処理、あんまりいいのが思い浮かばない
-# annoyで近傍を持ってきてシリーズを類推する
-annoy_db = AnnoyIndex(768, metric='euclidean') # shape、ハードエンコーディングだけどしらね
-annoy_db.load('./features/annoy_db.ann')
-# ベクトルvを与えると、近傍n個のアイテムを取り出せる
-# include_distancesはTrueで２地点間の距離を含める
-train_embeddings = np.load('./features/platform_genre_name_train_sentence_vectors.npy')
-test_embeddings  = np.load('./features/platform_genre_name_test_sentence_vectors.npy')
-train_test_embeddings = np.concatenate([train_embeddings, test_embeddings], axis=0)
-for i, embeddings in enumerate(train_test_embeddings):
-    nn_idxs, distances = annoy_db.get_nns_by_vector(embeddings, 11, search_k=-1, include_distances=True)
-    print(nn_idxs)
-    nn_names = train_test.iloc[nn_idxs]['Name'].values.tolist()
-    for j, (name, dis) in enumerate(zip(nn_names, distances)):
-        print(j, dis, name)
-    # print(train_test.iloc[nn_idxs]['Name'])
-    # vec = CountVectorizer(ngram_range=(2, 2)).fit(nn_names)
-    # bag_of_words = vec.transform(nn_names)
-    # sum_words = bag_of_words.sum(axis=0)
-    # print(sum_words)
-    # words_freq = [(word, sum_words[0, idx]) for word, idx in vec.vocabulary_.items() if sum_words[0, idx] > 1]
-    # words_freq =sorted(words_freq, key = lambda x: x[1], reverse=False)
-    # print(words_freq)
-    # # print(vec)
-    
-    exit()
-'''
-
-
-
-
-
-
-
-# # 連続するN単語を頻出順に表示する＆出現数を特徴量にする # NOTE: そんなに効いてない気がする(消すこともないっちゃないんだけど...)
-# def get_top_text_ngrams(corpus, n, g , s):
-#     vec = CountVectorizer(ngram_range=(g, g)).fit(corpus)
-#     bag_of_words = vec.transform(corpus)
-#     sum_words = bag_of_words.sum(axis=0) 
-#     words_freq = [(word, sum_words[0, idx]) for word, idx in vec.vocabulary_.items() if sum_words[0, idx] > s]
-#     words_freq =sorted(words_freq, key = lambda x: x[1], reverse=False)
-#     return words_freq[:n]
-
-# most_common_bi = get_top_text_ngrams(train_test.Name,10000,2,5)
-# most_common_bi = dict(most_common_bi)
-
-# train_test["num_Series"] = 0
-# for i in most_common_bi:
-#     idx = train_test[train_test["Name"].str.contains(i)].index
-#     train_test.iloc[idx, -1] = most_common_bi[i]
-
-
-print(train_test.head())
-print(train_test.columns.tolist())
-# exit()
-
 lgbm_params = {
     'objective': 'rmse', # 目的関数. これの意味で最小となるようなパラメータを探します. 
     'learning_rate': 0.1, # 学習率. 小さいほどなめらかな決定境界が作られて性能向上に繋がる場合が多いです、がそれだけ木を作るため学習に時間がかかります
@@ -358,16 +288,11 @@ cab_params = {
     'depth': 5,
 }
 
-
 # trainとtestに分割
 train, test = train_test[:train_length], train_test[train_length:]
 
 y = train['Global_Sales']
 y = np.log1p(y) # log + 1 変換
-
-print(y)
-print(train)
-print(test)
 
 # 使えなさそうなドロップするカラム TODO: ここobjectの列を列挙するように変えてもいいと思ったけど、Salesがありましたね...
 drop_column = ['Name','Platform','Genre','Publisher','NA_Sales','EU_Sales','JP_Sales','Other_Sales','Global_Sales','Developer','Rating','Platform_and_Genre','Platform_and_Genre_and_Binning_Year']
@@ -393,13 +318,11 @@ test = test.drop(drop_column, axis=1)
 
 # training data の target と同じだけのゼロ配列を用意
 # float にしないと悲しい事件が起こるのでそこだけ注意
-cab_oof_pred = np.zeros_like(y, dtype=np.float)
+knn_oof_pred  = np.zeros_like(y, dtype=np.float)
+cab_oof_pred  = np.zeros_like(y, dtype=np.float)
 lgbm_oof_pred = np.zeros_like(y, dtype=np.float)
 scores, models = [], []
 skf = StratifiedKFold(n_splits=N_SPLITS, random_state=RANDOM_SEED, shuffle=True)
-# num_bins = np.int(1 + np.log2(len(train)))
-# bins = pd.cut(train['Global_Sales'], bins=num_bins, labels=False)
-# for i, (train_idx, valid_idx) in enumerate(skf.split(train, bins.values)):
 for i, (train_idx, valid_idx) in enumerate(skf.split(train, train['Publisher'])):
     x_train, x_valid = train.iloc[train_idx], train.iloc[valid_idx]
     y_train, y_valid = y.iloc[train_idx], y.iloc[valid_idx]
@@ -465,6 +388,7 @@ sub_df.to_csv(f'./submission/cv:{score}_sub.csv', index=False)
 # feature importanceの可視化
 feature_importance_df = pd.DataFrame()
 for i, model in enumerate(models):
+    if i%2==0: continue
     _df = pd.DataFrame()
     _df['feature_importance'] = model.feature_importances_
     _df['column'] = train.drop(drop_column, axis=1).columns
